requirements_metrics.txt
# Dependencies for the metrics tracking system
pandas>=1.3.0
tabulate>=0.8.9
numpy>=1.20.0
matplotlib>=3.4.0  # For potential visualization
statistics>=1.0.3.5
google-cloud-speech
google-genai
